Fixing issue 154 for branch 0_7_EA
git-svn-id: https://svn.apache.org/repos/asf/incubator/syncope/branches/0_7_EA@1247107 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/syncope/core/rest/data/AbstractAttributableDataBinder.java
@@ -319,10 +319,24 @@ public abstract class AbstractAttributableDataBinder {
                 if (attribute == null) {
                     LOG.debug("No attribute found for schema {}", schema);
                 } else {
-                    attributable.removeAttribute(attribute);
+                    String newValue = null;
+                    for (AttributeMod mod : attributableMod.
+                            getAttributesToBeUpdated()) {
 
-                    attributeDAO.delete(attribute.getId(),
-                            attributableUtil.attributeClass());
+                        if (schema.getName().equals(mod.getSchema())) {
+                            newValue = mod.getValuesToBeAdded().
+                                    iterator().next();
+                        }
+                    }
+
+                    if (!schema.isUniqueConstraint()
+                            || (!attribute.getUniqueValue().getStringValue().
+                            equals(newValue))) {
+
+                        attributable.removeAttribute(attribute);
+                        attributeDAO.delete(attribute.getId(),
+                                attributableUtil.attributeClass());
+                    }
                 }
 
                 for (SchemaMapping mapping : resourceDAO.findAllMappings()) {
